requirements.txt
flask
sqlalchemy
pymysql
py-emails
pyjwt
Pillow